use pointer for methods
