--- .github/cloud-samples-tools/pkg/utils/packages.go
@@ -1,49 +1,0 @@
-/*
- Copyright 2024 Google LLC
-
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at
-
-      https://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
-*/
-
-package utils
-
-import (
-	"io/fs"
-	"os"
-	"slices"
-)
-
-// FindAllPackages finds all the packages in the given root directory.
-func FindAllPackages(root string, config Config) ([]string, error) {
-	var packages []string
-	err := fs.WalkDir(os.DirFS(root), ".",
-		func(path string, d os.DirEntry, err error) error {
-			if err != nil {
-				return err
-			}
-			if path == "." {
-				return nil
-			}
-			if slices.Contains(config.ExcludePackages, path) {
-				return nil
-			}
-			if d.IsDir() && config.Matches(path) && config.IsPackageDir(path) {
-				packages = append(packages, path)
-				return nil
-			}
-			return nil
-		})
-	if err != nil {
-		return []string{}, err
-	}
-	return packages, nil
-}
