+++ .github/cloud-samples-tools/test/config_test.go
@@ -1,7 +1,7 @@
 package test
 
 import (
-	"cloud-samples-tools/pkg/utils"
+	c "cloud-samples-tools/pkg/config"
 	"os"
 	"path/filepath"
 	"testing"
@@ -9,15 +9,15 @@ import (
 
 func TestLoadConfigEmpty(t *testing.T) {
 	path := filepath.Join("testdata", "config", "empty.json")
-	_, err := utils.LoadConfig(path)
+	_, err := c.LoadConfig(path)
 	expectError(t, err)
 }
 
 func TestLoadConfigDefaultValues(t *testing.T) {
 	path := filepath.Join("testdata", "config", "default-values.json")
-	config, err := utils.LoadConfig(path)
+	config, err := c.LoadConfig(path)
 	ok(t, err)
-	expected := utils.Config{
+	expected := c.Config{
 		PackageFile: []string{"package.json"},
 		Match:       []string{"*"},
 	}
@@ -26,9 +26,9 @@ func TestLoadConfigDefaultValues(t *testing.T) {
 
 func TestLoadConfigComments(t *testing.T) {
 	path := filepath.Join("testdata", "config", "comments.jsonc")
-	config, err := utils.LoadConfig(path)
+	config, err := c.LoadConfig(path)
 	ok(t, err)
-	expected := utils.Config{
+	expected := c.Config{
 		PackageFile: []string{"package.json"},
 		Match:       []string{"*"},
 	}
@@ -40,16 +40,16 @@ func TestSaveLoadConfig(t *testing.T) {
 	ok(t, err)
 	defer os.Remove(file.Name())
 
-	config := utils.Config{
+	config := c.Config{
 		PackageFile:     []string{"package.json"},
 		Ignore:          []string{"node_modules/", "*.md"},
 		Match:           []string{"*.js"},
 		ExcludePackages: []string{"excluded"},
 	}
-	err = utils.SaveConfig(config, file)
+	err = config.Save(file)
 	ok(t, err)
 
-	loadedConfig, err := utils.LoadConfig(file.Name())
+	loadedConfig, err := c.LoadConfig(file.Name())
 	ok(t, err)
 
 	equals(t, config, loadedConfig)
@@ -58,41 +58,77 @@ func TestSaveLoadConfig(t *testing.T) {
 func TestMatchNoPatterns(t *testing.T) {
 	patterns := []string{}
 	path := "path/to/file.js"
-	assert(t, !utils.Match(patterns, path), "match(%v, %v) = false")
+	assert(t, !c.Match(patterns, path), "match(%v, %v) = false")
 }
 
 func TestMatchFilePattern(t *testing.T) {
 	patterns := []string{"*.js"}
 	path := "path/to/file.js"
-	assert(t, utils.Match(patterns, path), "match(%v, %v) = true")
+	assert(t, c.Match(patterns, path), "match(%v, %v) = true")
 }
 
 func TestMatchFilePath(t *testing.T) {
 	patterns := []string{"path/to/"}
 	path := "path/to/file.js"
-	assert(t, utils.Match(patterns, path), "match(%v, %v) = true")
+	assert(t, c.Match(patterns, path), "match(%v, %v) = true")
 }
 
 func TestIsPackageDirNotExists(t *testing.T) {
-	config := utils.Config{PackageFile: []string{"package.json"}}
+	config := c.Config{PackageFile: []string{"package.json"}}
 	dir := "path-does-not-exist"
 	assert(t, !config.IsPackageDir(dir), "config.IsPackageDir(%v) = false")
 }
 
 func TestIsPackageDirExists(t *testing.T) {
-	config := utils.Config{PackageFile: []string{"package.json"}}
+	config := c.Config{PackageFile: []string{"package.json"}}
 	dir := filepath.Join("testdata", "my-package")
 	assert(t, config.IsPackageDir(dir), "config.IsPackageDir(%v) = true")
 }
 
 func TestFindPackageRoot(t *testing.T) {
-	config := utils.Config{PackageFile: []string{"package.json"}}
+	config := c.Config{PackageFile: []string{"package.json"}}
 	path := filepath.Join("testdata", "my-file.txt")
 	equals(t, ".", config.FindPackage(path))
 }
 
 func TestFindPackageSubpackage(t *testing.T) {
-	config := utils.Config{PackageFile: []string{"package.json"}}
+	config := c.Config{PackageFile: []string{"package.json"}}
 	path := filepath.Join("testdata", "my-package", "subpackage", "my-file.txt")
 	equals(t, "testdata/my-package/subpackage", config.FindPackage(path))
+}
+
+func TestChangedRoot(t *testing.T) {
+	config := c.Config{PackageFile: []string{"package.json"}}
+	diffs := []string{
+		filepath.Join("testdata", "file.txt"),
+	}
+	got := config.Changed(diffs)
+	expected := []string{"."}
+	equals(t, expected, got)
+}
+
+func TestChangedPackage(t *testing.T) {
+	config := c.Config{
+		PackageFile: []string{"package.json"},
+		Match:       []string{"*"},
+	}
+	diffs := []string{
+		filepath.Join("testdata", "my-package", "file.txt"),
+	}
+	got := config.Changed(diffs)
+	expected := []string{"testdata/my-package"}
+	equals(t, expected, got)
+}
+
+func TestChangedSubpackage(t *testing.T) {
+	config := c.Config{
+		PackageFile: []string{"package.json"},
+		Match:       []string{"*"},
+	}
+	diffs := []string{
+		filepath.Join("testdata", "my-package", "subpackage", "file.txt"),
+	}
+	got := config.Changed(diffs)
+	expected := []string{"testdata/my-package/subpackage"}
+	equals(t, expected, got)
 }
